+++ ormconfig.json
@@ -5,9 +5,9 @@
     "username": "docker",
     "password": "ignite",
     "database": "rentx", 
-    "migrations": ["./src/database/migrations/*.ts"],
-    "entities": ["./src/modules/**/entities/*.ts"],
+    "migrations": ["./src/shared/infra/typeorm/migrations/*.ts"],
+    "entities": ["./src/modules/**/infra/typeorm/entities/*.ts"],
     "cli": {
-      "migrationsDir": "./src/database/migrations"
+      "migrationsDir": "./src/shared/infra/typeorm/migrations"
     }
   }
