+++ src/modules/accounts/useCases/createUser/CreateUserUseCase.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from "tsyringe";
 import { hash } from "bcryptjs";
-import { AppError } from "../../../../errors/AppError";
+import { AppError } from "../../../../shared/errors/AppError";
 import { IUsersRepository } from "../../repositories/IUsersRepository";
 import { ICreateUserDTO } from "../../dtos/ICreateUserDTO"
