+++ src/modules/cars/infra/typeorm/entities/Specification.ts
