+++ src/modules/cars/infra/typeorm/repositories/SpecificationRepository.ts
@@ -1,6 +1,6 @@
 import { getRepository, Repository } from "typeorm";
-import { Specification } from "../../entities/Specification";
-import { ICreateSpecificationDTO, ISpecificationRepository } from "../ISpecificationsRepository";
+import { Specification } from "../entities/Specification";
+import { ICreateSpecificationDTO, ISpecificationRepository } from "../../../repositories/ISpecificationsRepository";
 
 class SpecificationRepository implements ISpecificationRepository{
     private repository: Repository<Specification>;
