+++ src/modules/cars/repositories/ICategoriesRepository.ts
@@ -1,4 +1,4 @@
-import {Category} from '../entities/Category'
+import {Category} from '../infra/typeorm/entities/Category'
 
 // Conceito DTO => Data transfer object
 interface ICategorieDTO {
