+++ src/modules/cars/useCases/createCategory/CreateCategoryUseCase.ts
@@ -1,5 +1,5 @@
 import { inject, injectable } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
+import { AppError } from "../../../../shared/errors/AppError";
 import { ICategoriesRepository } from '../../repositories/ICategoriesRepository';
 
 interface IRequest {
