+++ src/modules/cars/useCases/listCategories/ListCategoriesUseCase.ts
@@ -1,5 +1,5 @@
 import { inject, injectable } from "tsyringe";
-import { Category } from '../../entities/Category';
+import { Category } from '../../infra/typeorm/entities/Category';
 import { ICategoriesRepository } from '../../repositories/ICategoriesRepository';
 
 @injectable()
