+++ src/shared/infra/http/middlewares/ensuredAuthenticated.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
 import { NextFunction } from "express";
 import { verify } from "jsonwebtoken";
-import { AppError } from "../errors/AppError"
-import { UsersRepository } from "../modules/accounts/repositories/implementations/UsersRepository";
+import { AppError } from "../../../errors/AppError"
+import { UsersRepository } from "../../../../modules/accounts/infra/typeorm/repositories/UsersRepository";
 
 interface IPayload {
     sub: string;
