+++ src/shared/infra/http/routes/specification.route.ts
@@ -1,7 +1,7 @@
 import {Router } from "express";
 
 import { ensureAuthenticated } from "../middlewares/ensuredAuthenticated";
-import { CreateSpecificationController } from "../modules/cars/useCases/createSpecification/CreateSpecificationController";
+import { CreateSpecificationController } from "../../../../modules/cars/useCases/createSpecification/CreateSpecificationController";
 
 const specificationsRoutes = Router();
