+++ src/shared/infra/http/routes/users.routes.ts
@@ -1,9 +1,9 @@
 import { Router } from "express";
 import multer from "multer";
 
-import uploadConfig from "../config/upload";
-import {CreateUserController} from "../modules/accounts/useCases/createUser/CreateUserController";
-import { UpdateUserAvatarController } from "../modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarController";
+import uploadConfig from "../../../../config/upload";
+import {CreateUserController} from "../../../../modules/accounts/useCases/createUser/CreateUserController";
+import { UpdateUserAvatarController } from "../../../../modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarController";
 import { ensureAuthenticated } from "../middlewares/ensuredAuthenticated";
 
 const usersRoutes = Router();
