+++ src/shared/infra/http/server.ts
@@ -2,13 +2,13 @@ import express, { NextFunction, Request, Response } from "express"
 import "express-async-errors";
 import swaggerUi from "swagger-ui-express"
 
-import "./database"
+import "../typeorm"
 
-import "./shared/container"
+import "../../container"
 
 import { router } from "./routes"
-import swaggerFile from "./swagger.json"
-import { AppError } from "./errors/AppError"
+import swaggerFile from "../../../swagger.json"
+import { AppError } from "../../errors/AppError"
 
 const app = express();
